+++ src/main/java/mcjty/rftoolsdim/dimensions/world/GenericWorldProvider.java
@@ -186,7 +186,7 @@ public class GenericWorldProvider extends WorldProvider implements  /*@todo impl
         if (dimensionInformation != null && dimensionInformation.getTerrainType() == TerrainType.TERRAIN_INVERTIGO) {
             for (int i = 0; i <= 15; ++i)
             {
-                float f1 = 1.0F - (float)i / 15.0F;
+                float f1 = 1.0F - i / 15.0F;
                 this.lightBrightnessTable[i] = (1.0F - f1) / (f1 * 3.0F + 1.0F) * 1.0F + 1.0F;
             }
             return;
